added and fixed tests and un-did the action cost Double conversion back to Long now

+++ src/main/kotlin/edu/unh/cs/ai/realtimesearch/planner/realtime/LssLrtaStarPlanner.kt
@@ -227,7 +227,7 @@ class LssLrtaStarPlanner<StateType : State<StateType>>(domain: Domain<StateType>
             }
 
             // Add the current state as the predecessor of the child state
-            successorNode.predecessors.add(Edge(node = sourceNode, action = successor.action, actionCost = successor.actionCost.toLong()))
+            successorNode.predecessors.add(Edge(node = sourceNode, action = successor.action, actionCost = successor.actionCost))
 
             // Skip if we got back to the parent
             if (successorState == sourceNode.parent.state) {
@@ -239,10 +239,10 @@ class LssLrtaStarPlanner<StateType : State<StateType>>(domain: Domain<StateType>
             if (successorNode.cost > successorGValueFromCurrent) {
                 // here we generate a state. We store it's g value and remember how to get here via the treePointers
                 successorNode.apply {
-                    cost = successorGValueFromCurrent.toLong()
+                    cost = successorGValueFromCurrent
                     parent = sourceNode
                     action = successor.action
-                    actionCost = successor.actionCost.toLong()
+                    actionCost = successor.actionCost
                 }
 
                 logger.debug { "Expanding from $sourceNode --> $successorState :: open list size: ${openList.size}" }
@@ -279,7 +279,7 @@ class LssLrtaStarPlanner<StateType : State<StateType>>(domain: Domain<StateType>
             val undiscoveredNode = Node(
                     state = successorState,
                     heuristic = domain.heuristic(successorState),
-                    actionCost = successor.actionCost.toLong(),
+                    actionCost = successor.actionCost,
                     action = successor.action,
                     parent = parent,
                     cost = MAX_VALUE,
