+++ src/main/kotlin/edu/unh/cs/ai/realtimesearch/planner/realtime/SZero.kt
@@ -239,7 +239,7 @@ class SZeroPlanner<StateType : State<StateType>>(domain: Domain<StateType>, conf
             }
 
             // Add the current state as the predecessor of the child state
-            successorNode.predecessors.add(SearchEdge(node = sourceNode, action = successor.action, actionCost = successor.actionCost.toLong()))
+            successorNode.predecessors.add(SearchEdge(node = sourceNode, action = successor.action, actionCost = successor.actionCost))
 
             // Skip if we got back to the parent
             if (successorState == sourceNode.parent.state) {
@@ -251,10 +251,10 @@ class SZeroPlanner<StateType : State<StateType>>(domain: Domain<StateType>, conf
             if (successorNode.cost > successorGValueFromCurrent) {
                 // here we generate a state. We store it's g value and remember how to get here via the treePointers
                 successorNode.apply {
-                    cost = successorGValueFromCurrent.toLong()
+                    cost = successorGValueFromCurrent
                     parent = sourceNode
                     action = successor.action
-                    actionCost = successor.actionCost.toLong()
+                    actionCost = successor.actionCost
                 }
 
                 logger.debug { "Expanding from $sourceNode --> $successorState :: open list size: ${openList.size}" }
@@ -290,7 +290,7 @@ class SZeroPlanner<StateType : State<StateType>>(domain: Domain<StateType>, conf
             val undiscoveredNode = Node(
                     state = successorState,
                     heuristic = domain.heuristic(successorState),
-                    actionCost = successor.actionCost.toLong(),
+                    actionCost = successor.actionCost,
                     action = successor.action,
                     parent = parent,
                     cost = MAX_VALUE,
